move validate_grammar() method to structured output backend
